codes/skill-visualizer/skill-converter/requirements.txt
pandas>=1.3.0
PyYAML>=5.4.1
numpy>=1.21.0